Fixed some spellings
