+++ client/src/Component/Hero.jsx
@@ -1,34 +1,35 @@
-import React from 'react';
-
 const Hero = () => {
-    const pralex =
-    {
-        /* The image used */
-        backgroundImage: `url('https://daisyui.com/images/stock/photo-1507358522600-9f71e620c44e.jpg')`,
+  const parallax = {
+    /* The image used */
+    backgroundImage: `url('https://daisyui.com/images/stock/photo-1507358522600-9f71e620c44e.jpg')`,
+    // TODO: This image doesn't go with the website theme.
 
-        /* Full height */
-        height: "100%",
+    /* Full height */
+    height: "100%",
 
-        /* Create the parallax scrolling effect */
-        backgroundAttachment: "fixed",
-        backgroundPosition: "center",
-        backgroundRepeat: "no-repeat",
-        backgroundSize: "cover",
-    }
+    /* Create the parallax scrolling effect */
+    backgroundAttachment: "fixed",
+    backgroundPosition: "center",
+    backgroundRepeat: "no-repeat",
+    backgroundSize: "cover",
+  };
 
-    return (
-        <div className="hero min-h-screen -z-40 " style={pralex}>
-            <div className="hero-overlay bg-opacity-60"></div>
-            <div className="hero-content text-center text-neutral-content z-0">
-                <div className="max-w-md">
-                    <h1 className="mb-5 text-5xl font-bold">Hello there</h1>
-                    <p className="mb-5">Provident cupiditate voluptatem et in. Quaerat fugiat ut assumenda excepturi exercitationem quasi. In deleniti eaque aut repudiandae et a id nisi.</p>
-                    <button className="btn btn-primary">Get Started</button>
-                </div>
-            </div>
+  return (
+    <div className="hero min-h-screen -z-40 " style={parallax}>
+      <div className="hero-overlay bg-opacity-60"></div>
+      <div className="hero-content text-center text-neutral-content z-0">
+        <div className="max-w-md">
+          <h1 className="mb-5 text-5xl font-bold">Hello there</h1>
+          <p className="mb-5">
+            Provident cupiditate voluptatem et in. Quaerat fugiat ut assumenda
+            excepturi exercitationem quasi. In deleniti eaque aut repudiandae et
+            a id nisi.
+          </p>
+          <button className="btn btn-primary">Get Started</button>
         </div>
-
-    );
+      </div>
+    </div>
+  );
 };
 
 export default Hero;
